Harden upgrade routines with proper type checks.

+++ includes/admin/migrations/3.0-general-options.php
@@ -4,7 +4,9 @@ defined( 'ABSPATH' ) or exit;
 
 // transfer option
 $options = get_option( 'mc4wp_lite' );
-update_option( 'mc4wp', $options );
+if( is_array( $options ) ) {
+	update_option( 'mc4wp', $options );
 
-// delete old option
-delete_option( 'mc4wp_lite' );
+	// delete old option
+	delete_option( 'mc4wp_lite' );
+}
